+++ wavelab/gui/baro_pressure_gui.py
@@ -23,7 +23,7 @@ GLOBAL_FIELDS = OrderedDict([
 LOCAL_FIELDS = OrderedDict([
     ('instrument_name', ['Instrument:', [
         'MS TruBlue 255', 'Onset Hobo U20', 'LevelTroll', 'RBRSolo',
-        'NOAA Station', 'Meso West'#, 'USGS Homebrew'
+        'NOAA Station', 'Meso West', 'VanEssen'#, 'USGS Homebrew'
         ], True]),
      ('stn_station_number', ['STN Site Id:', '']),
     ('stn_instrument_id', ['STN Instrument Id:', '']),