+++ CHANGELOG.md
@@ -9,11 +9,11 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ### Added 
 
--
+- 
 
 ### Changed  
 
--
+- 
 
 ### Deprecated 
 
@@ -22,13 +22,23 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 ### Removed 
 
 - 
-
 ### Fixed  
 
 - 
 
 ### Security  
 
-- 
+-
+## [v1.2.0](https://code.usgs.gov/wavelab/wavelab/-/tags/v1.2.0) - 2022-08-24
 
+### Added 
+
+- Added Van Essen to the list of possible baro sensors
+- Added software version statement to main GUI and all graphs
+- Added software version to metadata of output netCDF files
+
+### Changed  
+
+- Changed all instances of `matplotlib.use()` with `warn=False` to `force=False`
+- Changed pyinstaller to version 4.10
 ## [v1.1.0](https://code.usgs.gov/wavelab/wavelab/-/tags/v1.1.0) - 2022-07-12